[FLINK-21171] Rename packages for request-reply protocol Proto messages
